update swap unit test
